review: CodeRabbit round — quick wins applied, two declined with reasons
Applied: cap_cents bigint; slog LogValuer resolve before redaction;
price rows bounded so CostCents cannot overflow; per-attempt migrate
timeout; server Read/Idle timeouts (WriteTimeout stays 0 for streams);
CI completed-state assertion made errexit-proof; go-plane
least-privilege permissions; plane-admin.sh input validation and a
Secret-to-credential annotation binding (a 409 for a different name now
refuses instead of silently reusing the token); test lint nits
(errcheck, noctx); runbook fence language.

Declined: imagePullPolicy IfNotPresent (Never is deliberate — a
side-loaded local tag must never fall back to pulling a squattable
docker.io name); atomic budget reservation (monthly caps are
soft-stops, same read-then-forward semantics as tomte-old's meter;
revisit with P4c approvals). Fresh-plane live re-verified: migrate,
govern, chat completed and ledgered, idempotent re-govern.

+++ plane/internal/db/migrations/00001_governance.sql
@@ -12,7 +12,7 @@ CREATE TABLE credential (
     name        text NOT NULL UNIQUE,
     token_hash  bytea NOT NULL UNIQUE, -- sha256 of the issued token; plaintext never stored
     -- Monthly caps (calendar month, UTC). NULL = no cap of that kind.
-    cap_cents   integer CHECK (cap_cents >= 0),
+    cap_cents   bigint CHECK (cap_cents >= 0),
     cap_tokens  bigint CHECK (cap_tokens >= 0),
     created_at  timestamptz NOT NULL DEFAULT now()
 );
